requirements.txt
Appium-Python-Client
pytest
requests
simplejson
xlsxwriter
xlrd
openpyxl
clipboard
pymysql
pytest-html
black
beautifulsoup4
selenium-requests
isort
pylint
pytest-rerunfailures
python-jenkins
pytest-mock
xmltodict
selenium